Only request column_info for the requested tables
Calling ->column_info( undef, '%', '%') requests column information for
all tables in all schemas; although this information is cached, it can
be extremely slow, especially when the Postgres server is not local.
Look up and cache the table information on a table-by-table basis,
instead of all at once, for considerale performance savings.

Note that the caching assumes that table names are case-sensitively
unique within a database.