Stop manually disabling the "UTF8" flag before executing queries
f9471ffb added this code to work around the newly added UTF-8 support in
DBD::Oracle 1.16.  It is not known what the specifics of the failure it
was attempting to avoid are -- the best guess is that Oracle failed when
handed invalid UTF8 codepoints.

However, blindly turning the "UTF8" flag off is no better, as it may or
may not result in UTF-8 encoded data.  The correct fix is to be
consistent about the data (bytes vs characters) that is passed to the
database.