On Postgres, force the `client_encoding` to `UTF8`
The DBD::Pg driver is only able to reliably pass non-ASCII characters
back and forth to the database if the `client_encoding` of the
connection is set to `UTF8`.  The default is to use the server's
encoding, which is UTF8; however, some installations may have
explicitly set the `client_encoding` to `SQL_ASCII` or other values.
This causes test failures, and applications fail to be able to insert
non-ASCII characters into Postgres `text` columns.

Force the `client_encoding` to `UTF8` at connection time.  This is
done via a DSN parameter; it can also be done using `SET
CLIENT_ENCODING=UTF8` after connecting, but that would also require
re-setting L<DBD::Pg/pg_enable_utf8-integer> to a magic value of `-1`
for the change to take effect.  Altering the DSN prior seems a less
"magic" solution.