Support bind values for searches
This is mainly to reduce hard parses and help the DB
re-use cached queries, especially Oracle. Improving
use of cached query plans should improve performance.

Provide a PREFER_BIND option to make it possible to
disable this new feature since it has to dynamically
evaluate and update queries and this may have
issues on complicated queries not yet tested.