Add test showing long values get re-added on update
Values are currently truncated to fit in the defined field
length after the comparison to check if the new value is
the same as the current value. So if you submit a value
bigger than the column size repeatedly, the field will
update each time rather than correctly detecting it
as the same value.