Truncate values before checking for changes
Previously values were truncated after being compared
to the current value, so a value longer than the column
size will always appear different from the current value
even if the same value is provided multiple times.

Truncate first so the check will compare the current value
with the version of the new value that will actually be put
in the database.