Change how DistinctQueryAndCount builds query to fix sorting
If order by is on columns from the main table then old query generated by
DistinctQueryAndCount was like this:

    SELECT main.*, COUNT(main.id) OVER () FROM (SELECT ... ORDER BY main...) main;

It works until it doesn't. Order by clause should be on outer most query.

Some evidence:

* SQL standard: ORDER BY command cannot be used in a subquery. Google
  ANSI SQL order by in subquery.

* mariadb explanation: https://mariadb.com/kb/en/why-is-order-by-in-a-from-subquery-ignored/

* SQL server: https://learn.microsoft.com/en-us/sql/t-sql/queries/select-order-by-clause-transact-sql?view=sql-server-ver16
  "The ORDER BY clause is not valid in views, ... and *subqueries*, unless ..."

Even in cases where ORDER BY is allowed in a subquery there is stated
guaranty that order will preserved in the outer query.

New query:

    SELECT main.*, COUNT(main.id) OVER () FROM (SELECT ...) main ORDER BY main...;

Note that in case of empty ORDER BY the query is still wrapped to make
sure the window function counts distinct records, eg:

    SELECT main.*, COUNT(main.id) OVER () FROM (SELECT DISTINCT ...) main;